Moved initialization hooks to corresponding file.
