+++ xlive/H2MOD/Modules/Startup/Startup.cpp
@@ -12,12 +12,14 @@
 #include "Util\Debug\Debug.h"
 #include "Util\hash.h"
 #include "Util\filesys.h"
+#include "Util\Hooks\Hook.h"
 
+#include "H2MOD\Modules\Updater\Updater.h"
 #include "H2MOD\Modules\Config\Config.h"
 #include "H2MOD\Modules\Tweaks\Tweaks.h"
 #include "H2MOD\Modules\Accounts\AccountLogin.h"
 #include "H2MOD\Modules\Accounts\Accounts.h"
-
+#include "H2MOD\Modules\Networking\Networking.h"
 
 
 #ifndef NO_TRACE
@@ -26,7 +28,6 @@ logger *h2mod_log = nullptr;
 logger *network_log = nullptr;
 logger *checksum_log = nullptr;
 #endif
-
 
 ProcessInfo game_info;
 
@@ -317,6 +318,344 @@ inline std::string prepareLogFileName(std::string logFileName) {
 	return filename;
 }
 
+#pragma region initialization hook
+template <typename T = void>
+static inline T *GetAddress(DWORD client, DWORD server = 0)
+{
+	return reinterpret_cast<T*>(H2BaseAddr + (H2IsDediServer ? server : client));
+}
+
+#pragma region func_wrappers
+void real_math_initialize()
+{
+	typedef int (real_math_initialize)();
+	auto real_math_initialize_impl = GetAddress<real_math_initialize>(0x000340d7);
+	real_math_initialize_impl();
+}
+
+void async_initialize()
+{
+	typedef int (async_initialize)();
+	auto async_initialize_impl = GetAddress<async_initialize>(0x00032ce5);
+	async_initialize_impl();
+}
+
+bool init_gfwl_gamestore()
+{
+	typedef char (init_gfwl_gamestore)();
+	auto init_gfwl_gamestore_impl = GetAddress<init_gfwl_gamestore>(0x00202f3e);
+	return init_gfwl_gamestore_impl();
+}
+// not sure if this is all it does
+HANDLE init_data_checksum_info()
+{
+	typedef HANDLE init_data_checksum_info();
+	auto init_data_checksum_info_impl = GetAddress<init_data_checksum_info>(0x000388d3);
+	return init_data_checksum_info_impl();
+}
+
+// returns memory
+void *runtime_state_init()
+{
+	typedef void *runtime_state_init();
+	auto runtime_state_init_impl = GetAddress<runtime_state_init>(0x00037ed5);
+	return runtime_state_init_impl();
+}
+
+void global_preferences_initialize()
+{
+	typedef void global_preferences_initialize();
+	auto global_preferences_initialize_impl = GetAddress<global_preferences_initialize>(0x325FD);
+	global_preferences_initialize_impl();
+}
+
+void font_initialize()
+{
+	typedef void __cdecl font_initialize();
+	auto font_initialize_impl = GetAddress<font_initialize>(0x00031dff);
+	font_initialize_impl();
+}
+
+bool tag_files_open()
+{
+	typedef bool tag_files_open();
+	auto tag_files_open_impl = GetAddress<tag_files_open>(0x30D58);
+	return tag_files_open_impl();
+}
+
+void init_timing(int a1)
+{
+	typedef DWORD(__cdecl init_timing)(int a1);
+	auto init_timing_impl = GetAddress<init_timing>(0x37E39);
+	init_timing_impl(a1);
+}
+
+void game_state_initialize(void *data)
+{
+	typedef void __fastcall game_state_initialize(void *data);
+	auto game_state_initialize_impl = GetAddress<game_state_initialize>(0x00030aa6);
+	game_state_initialize_impl(data);
+}
+
+bool rasterizer_initialize()
+{
+	typedef char rasterizer_initialize();
+	auto rasterizer_initialize_impl = GetAddress<rasterizer_initialize>(0x00263359);
+	return rasterizer_initialize_impl();
+}
+
+bool input_initialize()
+{
+	typedef char input_initialize();
+	auto input_initialize_impl = GetAddress<input_initialize>(0x2FD23);
+	return input_initialize_impl();
+}
+
+void sound_initialize()
+{
+	typedef void sound_initialize();
+	auto sound_initialize_impl = GetAddress<sound_initialize>(0x2979E);
+	return sound_initialize_impl();
+}
+
+#pragma endregion
+
+int flag_log_count[flags::count];
+BOOL __cdecl is_init_flag_set(flags id)
+{
+	if (flag_log_count[id] < 10)
+	{
+		TRACE_GAME("is_init_flag_set() : flag %i", id);
+		flag_log_count[id]++;
+		if (flag_log_count[id] == 10)
+			TRACE_GAME("is_init_flag_set() : flag %i logged to many times ignoring", id);
+	}
+	DWORD* init_flags_array = reinterpret_cast<DWORD*>(H2BaseAddr + 0x0046d820);
+	return init_flags_array[id] != 0;
+}
+
+const static int max_mointor_count = 9;
+bool engine_basic_init()
+{
+	DWORD* flags_array = reinterpret_cast<DWORD*>(H2BaseAddr + 0x0046d820);
+	memset(flags_array, 0x00, flags::count); // should be zero initalized anyways but the game does it
+
+	flags_array[flags::nointro] = H2Config_skip_intro;
+
+	HANDLE(*fn_c000285fd)() = (HANDLE(*)())(GetAddress(0x000285fd));
+
+	init_gfwl_gamestore();
+	init_data_checksum_info();
+	runtime_state_init();
+
+	int arg_count;
+	wchar_t **cmd_line_args = LOG_CHECK(CommandLineToArgvW(GetCommandLineW(), &arg_count));
+	if (cmd_line_args && arg_count > 1) {
+		for (int i = 1; i < arg_count; i++) {
+			wchar_t* cmd_line_arg = cmd_line_args[i];
+
+			if (wcsicmp(cmd_line_arg, L"-windowed") == 0) {
+				flags_array[flags::windowed] = 1;
+			}
+			else if (wcsicmp(cmd_line_arg, L"-nosound") == 0) {
+				flags_array[flags::nosound] = 1;
+				WriteValue(H2BaseAddr + 0x479EDC, 1);
+			}
+			else if (wcsicmp(cmd_line_arg, L"-novsync") == 0) {
+				flags_array[flags::novsync] = 1;
+			}
+			else if (wcsicmp(cmd_line_arg, L"-nointro") == 0) {
+				flags_array[flags::nointro] = 1;
+			}
+			else if (wcsnicmp(cmd_line_arg, L"-monitor:", 9) == 0) {
+				int monitor_id = _wtol(&cmd_line_arg[9]);
+				flags_array[flags::monitor_count] = min(max(0, monitor_id), max_mointor_count);
+			}
+			else if (wcsicmp(cmd_line_arg, L"-highquality") == 0) {
+				flags_array[flags::high_quality] = 1;
+			}
+			else if (wcsicmp(cmd_line_arg, L"-depthbiasfix") == 0)
+			{
+				// Fixes issue #118
+				/* g_depth_bias always NULL rather than taking any value from
+				shader tag before calling g_D3DDevice->SetRenderStatus(D3DRS_DEPTHBIAS, g_depth_bias); */
+				NopFill<8>(reinterpret_cast<DWORD>(GetAddress(0x269FD5)));
+			}
+#ifdef _DEBUG
+			else if (wcsnicmp(cmd_line_arg, L"-dev_flag:", 10) == 0) {
+				int flag_id = _wtol(&cmd_line_arg[10]);
+				flags_array[min(max(0, flag_id), flags::count - 1)] = 1;
+			}
+#endif
+		}
+	}
+	LocalFree(cmd_line_args);
+
+	if (flags_array[flags::unk22])
+		init_timing(1000 * flags_array[flags::unk22]);
+	real_math_initialize();
+	async_initialize();
+	global_preferences_initialize();
+	font_initialize();
+
+	if (!LOG_CHECK(tag_files_open()))
+		return false;
+	void *var_c004ae8e0 = GetAddress(0x004ae8e0);
+	game_state_initialize(var_c004ae8e0);
+
+	// modifies esi need to check what the caller sets that too
+	//char(*fn_c001a9de6)() = (char(*)())(GetAddress(0x001a9de6));
+	//char result_c001a9de6 = fn_c001a9de6();
+	if (!LOG_CHECK(rasterizer_initialize()))
+		return false;
+
+	input_initialize();
+	// not needed but bungie did it so there might be some reason
+	sound_initialize();
+
+	struct FakePBuffer {
+		HANDLE id;
+		DWORD dwSize;
+		DWORD magic;
+		LPBYTE pbData;
+	};
+	//extern LONG WINAPI XLivePBufferAllocate(DWORD size, FakePBuffer **pBuffer);
+	//extern DWORD WINAPI XLivePBufferSetByte(FakePBuffer * pBuffer, DWORD offset, BYTE value);
+	LONG(__stdcall* XLivePBufferAllocate)(DWORD size, FakePBuffer **pBuffer) = (LONG(__stdcall*)(DWORD, FakePBuffer**))(GetAddress(0x0000e886));
+	DWORD(__stdcall* XLivePBufferSetByte)(FakePBuffer * pBuffer, DWORD offset, BYTE value) = (DWORD(__stdcall*)(FakePBuffer*, DWORD, BYTE))(GetAddress(0x0000e880));
+	DWORD* var_c00479e78 = GetAddress<DWORD>(0x00479e78);
+	XLivePBufferAllocate(2, (FakePBuffer**)&var_c00479e78);
+	XLivePBufferSetByte((FakePBuffer*)var_c00479e78, 0, 0);
+	XLivePBufferSetByte((FakePBuffer*)var_c00479e78, 1, 0);
+
+	//SLDLInitialize
+
+	// does some weird stuff with sound, might setup volume
+	HANDLE result_c000285fd = fn_c000285fd();
+
+	//SLDLOpen
+	//SLDLConsumeRight
+	//SLDLClose
+	// This call would disable the multiplayer buttons in the mainmenu. Likely setup this way from SLDL.
+	//XLivePBufferSetByte((FakePBuffer*)var_c00479e78, 0, 1);
+
+	return true;
+}
+
+#pragma region func_wrappers
+bool InitPCCInfo()
+{
+	typedef bool __cdecl InitPCCInfo();
+	auto InitPCCInfoImpl = GetAddress<InitPCCInfo>(0x260DDD);
+	return InitPCCInfoImpl();
+}
+
+void run_main_loop()
+{
+	typedef int __cdecl run_main_loop();
+	auto run_main_loop_impl = GetAddress<run_main_loop>(0x39E2C);
+	run_main_loop_impl();
+}
+
+void main_engine_dispose()
+{
+	typedef int main_engine_dispose();
+	auto main_engine_dispose_impl = GetAddress<main_engine_dispose>(0x48A9);
+	main_engine_dispose_impl();
+}
+
+void show_error_message_by_id(int id)
+{
+	typedef void __cdecl show_error_message_by_id(int id);
+	auto show_error_message_by_id_impl = GetAddress<show_error_message_by_id>(0x4A2E);
+	show_error_message_by_id_impl(id);
+}
+#pragma endregion
+
+void show_fatal_error(int error_id)
+{
+	TRACE_FUNC("error_id : %d", error_id);
+
+	auto destory_window = [](HWND handle) {
+		if (handle)
+			DestroyWindow(handle);
+	};
+
+	HWND hWnd = *GetAddress<HWND>(0x46D9C4);
+	HWND d3d_window = *GetAddress<HWND>(0x46D9C8); // not sure what this window is actual for, used in IDirect3DDevice9::Present
+	destory_window(hWnd);
+	destory_window(d3d_window);
+	show_error_message_by_id(error_id);
+}
+
+char is_remote_desktop()
+{
+	TRACE_FUNC("check disabled");
+	return 0;
+}
+
+
+int __stdcall WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdLine, int nShowCmd)
+{
+	// set args
+	WriteValue(H2BaseAddr + 0x46D9BC, lpCmdLine); // command_line_args
+	WriteValue(H2BaseAddr + 0x46D9C0, hInstance); // g_instance
+	WriteValue(H2BaseAddr + 0x46D9CC, nShowCmd); // g_CmdShow
+
+	// window setup
+	wcscpy_s(reinterpret_cast<wchar_t*>(H2BaseAddr + 0x46D9D4), 0x40, L"halo"); // ClassName
+	wcscpy_s(reinterpret_cast<wchar_t*>(H2BaseAddr + 0x46DA54), 0x40, L"Halo 2 - Project Cartographer"); // WindowName
+	WNDPROC g_WndProc = reinterpret_cast<WNDPROC>(H2BaseAddr + 0x790E);
+	WriteValue(H2BaseAddr + 0x46D9D0, g_WndProc); // g_WndProc_ptr
+	if (!LOG_CHECK(InitPCCInfo()))
+	{
+		TRACE_FUNC("Failed to get PCC info / insufficient system resources");
+		run_async(
+			MessageBoxA(NULL, "Failed to get compatibility info.", "PCC Error", S_OK);
+		)
+			show_error_message_by_id(108);
+		// todo: load some default values here?
+	}
+	// mouse cursor setup
+	HCURSOR cursor = LOG_CHECK(LoadCursor(NULL, MAKEINTRESOURCE(0x7F00)));
+	WriteValue(H2BaseAddr + 0x46D9B8, cursor); // g_hCursor
+
+	// mess around with xlive (not calling XLiveInitialize etc)
+	WriteValue<BYTE>(H2BaseAddr + 0x4FAD98, 1);
+
+	// intialize some basic game subsystems
+	if (LOG_CHECK(engine_basic_init()))
+	{
+		run_main_loop(); // actually run game
+		main_engine_dispose(); // cleanup
+	}
+	else
+	{
+		TRACE_FUNC("Engine startup failed!");
+		show_fatal_error(108);
+		return 1;
+	}
+
+	int g_fatal_error_id = *reinterpret_cast<int*>(H2BaseAddr + 0x46DAD4);
+	if (g_fatal_error_id) // check if the game exited cleanly
+	{
+		show_fatal_error(g_fatal_error_id);
+		return 1;
+	}
+	else
+	{
+		return 0;
+	}
+}
+
+void applyGameInitializationHooks()
+{
+	WriteJmpTo(GetAddress(0x7E43), WinMain);
+	WriteJmpTo(GetAddress(0x39EA2), is_remote_desktop);
+	WriteJmpTo(GetAddress(0x4544), is_init_flag_set);
+}
+#pragma endregion
+
 ///Before the game window appears
 void InitH2Startup() {
 	InitializeCriticalSection(&log_section);
@@ -381,6 +720,7 @@ void InitH2Startup() {
 	}
 
 	InitH2Config();
+
 #ifndef NO_TRACE
 	EnterCriticalSection(&log_section);
 	if (H2Config_debug_log) {
@@ -394,8 +734,10 @@ void InitH2Startup() {
 	checksum_log = logger::create(prepareLogFileName("checksum"), true);
 	LeaveCriticalSection(&log_section);
 #endif
+
 	InitH2Accounts();
 
+	applyGameInitializationHooks();
 	configureXinput();
 
 	InitH2Tweaks();
@@ -407,8 +749,6 @@ void InitH2Startup() {
 	initGSRunLoop();
 	addDebugText("ProcessStartup finished.");
 }
-
-#include "H2MOD\Modules\Updater\Updater.h"
 
 ///After the game window appears
 void InitH2Startup2() {
